refactor : 트랜잭션 처리를 위해 메서드를 별도 서비스 클래스로 분리 - Runnable 반환 구조로 인해 @Transactional이 적용되지 않던 메서드 분리 - 프록시 기반 트랜잭션 처리를 위해 서비스 클래스로 이동
